LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-position-grid-001-expected.txt

1
2
3

FAIL .target 1 assert_equals:
<div class="target target1-l" data-offset-x="100" data-expected-height="100"></div>
offsetLeft expected 100 but got 200
FAIL .target 2 assert_equals:
<div class="target target1-r" data-offset-x="194" data-expected-height="100"></div>
offsetLeft expected 194 but got 294
FAIL .target 3 assert_equals:
<div class="target target1-t" data-offset-y="50" data-expected-width="100"></div>
offsetTop expected 50 but got 100
FAIL .target 4 assert_equals:
<div class="target target1-b" data-offset-y="145" data-expected-width="100"></div>
offsetTop expected 145 but got 195
